requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
httpx
Pillow
torch 
torchvision
